move unitgroup and formatter into experimental package; make DefaultFormatter a class
