+++ project/Build.scala
@@ -169,7 +169,8 @@ object Console {
          squants.space._,
          squants.thermal._,
          squants.time._,
-         squants.unitgroups.UnitGroup,
+         squants.experimental.formatter._,
+         squants.experimental.unitgroups.UnitGroup,
          squants.DimensionlessConversions._,
          squants.electro.CapacitanceConversions._,
          squants.electro.ConductivityConversions._,
