+++ shared/src/main/scala/squants/experimental/unitgroups/UnitGroup.scala
@@ -1,4 +1,4 @@
-package squants.unitgroups
+package squants.experimental.unitgroups
 
 import scala.collection.immutable.{Set, SortedSet}
