+++ shared/src/main/scala/squants/experimental/unitgroups/uscustomary/package.scala
@@ -1,4 +1,4 @@
-package squants.unitgroups
+package squants.experimental.unitgroups
 
 import squants.UnitOfMeasure
 import squants.mass._
