+++ shared/src/test/scala/squants/experimental/formatter/DefaultFormatterSpec.scala
@@ -1,4 +1,4 @@
-package squants.formatter
+package squants.experimental.formatter
 
 import org.scalatest.{FlatSpec, Matchers}
 import squants.UnitOfMeasure
@@ -6,12 +6,12 @@ import squants.mass.{Kilograms, Mass}
 import squants.mass.MassConversions._
 import squants.space.{Inches, Length, Yards, UsMiles}
 import squants.space.LengthConversions._
-import squants.unitgroups.UnitGroup
-import squants.unitgroups.uscustomary.space.UsCustomaryLengths
+import squants.experimental.unitgroups.UnitGroup
+import squants.experimental.unitgroups.uscustomary.space.UsCustomaryLengths
 
 class DefaultFormatterSpec extends FlatSpec with Matchers {
 
-  val usLengthFormatter = new DefaultFormatter[Length] { val unitGroup = UsCustomaryLengths }
+  val usLengthFormatter = new DefaultFormatter(UsCustomaryLengths)
 
   behavior of "DefaultFormatter"
 
@@ -26,7 +26,7 @@ class DefaultFormatterSpec extends FlatSpec with Matchers {
 
   it should "use the input quantity's unit when the unitgroup set is empty" in {
     val emptyUnitGroup = new UnitGroup[Mass] { val units = Set.empty[UnitOfMeasure[Mass]] }
-    val emptyUnitFormatter = new DefaultFormatter[Mass] { val unitGroup = emptyUnitGroup }
+    val emptyUnitFormatter = new DefaultFormatter(emptyUnitGroup)
 
     emptyUnitFormatter.inBestUnit(500.kg).unit should be (Kilograms)
   }
